Forgot to add the data dump + moved all model classes into __init__.py, therefore also changed import statements in the controllers
