--- application/data/models/auto_order.py
@@ -1,6 +1,0 @@
-from sqlalchemy import Integer, Column, Date, Numeric, ForeignKey
-
-from application.data.db import Base
-
-
-
