--- application/data/models/car_model.py
@@ -1,9 +1,0 @@
-from sqlalchemy import Integer, String, Column
-from sqlalchemy.orm import relationship
-
-from application.data.db import Base
-
-from application.data.models.customer_car import CustomerCar
-
-
-
