--- application/data/models/car_model_spare_part.py
@@ -1,9 +1,0 @@
-from sqlalchemy import Integer, Column, ForeignKey
-from sqlalchemy.orm import relationship
-
-from application.data.db import Base
-
-from application.data.models.car_model import CarModel
-
-
-
