--- application/data/models/company.py
@@ -1,8 +1,0 @@
-from sqlalchemy import Integer, String, Column, ForeignKey
-from sqlalchemy.orm import relationship
-
-from application.data.db import Base
-
-
-
-
