--- application/data/models/customer.py
@@ -1,10 +1,0 @@
-from sqlalchemy import Integer, String, Column, Boolean
-from sqlalchemy.orm import relationship
-
-from application.data.db import Base
-
-from application.data.models.customer_order import CustomerOrder
-from application.data.models.customer_address import CustomerAddress
-
-
-
