--- application/data/models/spare_part.py
@@ -1,20 +1,0 @@
-from sqlalchemy import Integer, String, Column, Float
-from sqlalchemy.orm import relationship
-
-from application.data.db import Base
-
-from application.data.models.car_model_spare_part import CarModelSparePart
-
-
-class SparePart(Base):
-    __tablename__ = "spare_parts"
-
-    product_number = Column(Integer, primary_key=True, nullable=False)
-    name = Column(String(100), nullable=False)
-    description = Column(String(45), nullable=False)
-    sell_price = Column(Float, nullable=False)
-    car_models = relationship("CarModelSparePart", back_populates="spare_parts")
-    customer_orders = relationship("OrderDetail", back_populates="spare_parts")
-    manufacturers = relationship("SparePartManufacturer", back_populates="spare_part")
-    suppliers = relationship("SparePartSupplier", back_populates="spare_part")
-    stores = relationship("SparePartInStore", back_populates="spare_parts")
