+++ application/data/repository/car_model_repository.py
@@ -1,5 +1,5 @@
 from application.data.db import session
-from application.data.models.car_model import CarModel
+from application.data.models import CarModel
 
 
 def create_car_model(car_model):
